backend/requirements.txt
redis
cantools
python-can
openpyxl
uvicorn
fastapi
